# workshop/product-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a Product API.",
        "title": "Product API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/product/{id}": {
            "get": {
                "description": "Retrieves a product's details using its ID.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get product detail by ID",
                "operationId": "get-product-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "product id=2 not found in system"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Product name 1"
                },
                "price": {
                    "type": "number",
                    "example": 100.5
                },
                "stock": {
                    "type": "integer",
                    "example": 10
                }
            }
        }
    }
}